chore: fix AshStub function matching for empty aggregates (#2290)
When calling an ash function from Javascript, the code tries to
determine a type for empty arrays/aggregates based on the expected
parameter type. It first matches the function with ANY type, and then
passes a typeHint to the ValueConverter to generate an empty typed
array/aggregate.

However, ANY type on the right hand side of a parameter match does not
match with anything else than an ANY parameter. Instead, replace any ANY
type values with the expected target type during matching, and replace
it with the hinted empty aggregate in the second run.

Furthermore, the previous implementation effectively does nothing, since
it did not used the values from the second run. Fix this, and improve
performance by not converting all parameters twice even when no ANY
types are used.