Mark typescript parameters as optional (#2284)
* feat: allow trailing undefined values to be passed to ash functions

Javascript allows callers to pass additional arguments to a function,
even if they were not declared in the function signature, or omit
parameters from a function call, which will make them have an
`undefined` value.
To accommodate this, Typescript allows parameters to be marked as
optional, which allows the caller to either omit a trailing parameter,
or explicitly pass `undefined` for that parameter.

We have several RuntimeLibrary functions where the last parameter is
optional, because there is an overladed version where all other
parameter are identical. While omitting parameters works in this case,
explicitly passing `undefined` does not.

To make this more convenient to use, and more compatible with
Javascript/Typescript convention, strip `undefined` parameters from the
end of the arguments array. Since undefined values are (so far) not at
all supported by either scripting environment, this should not pose any
incompatiblity issues.

* chore(TypescriptDefinition): use class to represent function/parameter

* feat(TypescriptDefinition): mark parameters as optional

Instead of printing two overloads that are fully equal except for one
additional parameter, merge the overloads and mark that parameter as
optional.

  function foo(a: string): boolean
  function foo(a: string, b: int): boolean
  function foo(a: string, b: int, c: int): boolean
  function foo(a: string, b: int, c: int, d: string): number

it will now print:

  function foo(a: string, b?: int): boolean
  function foo(a: string, b: int, c?: int): boolean
  function foo(a: string, b: int, c: int, d: string): number

NOTE: Marking multiple consecutive paramters optional is not correct,
because otherwise TypeScript will allow to pass `undefined` for a
"middle" parameter:

  foo(someA, undefined, someC);

This would be incorrect and lead to a function not found error.