Refresh modifiers when lifting Standard restrictions.
When liberating the king or breaking ronin, previously permed skills that
are too old for Standard restrictions will be available again.

I'm not actually sure off-hand whether Mafia knows a priori if those skills
are permed but unusable, or if we need to fetch the charsheet first. I know
there have been occasional issues with KoL not showing them in the past.