Throw exceptions for nested null / undefined.
It seems that NativeArray / NativeObject convert undefined -> null, so we
cannot differentiate at that level. It doesn't bother me too much. Both can
cause problems in ASH.

Note that in JavaScriptCommandTest, we cannot specify an object literal
directly -- this is one of the scenarios outlined in Gary Bernhardt's "wat"
talk, where the opening brace is parsed as the beginning of a block scope
and not the first character of an object literal. As such, those tests
begin with "o = " to provide the correct context to the interpreter.