Test that null / undefined can be used in pure JS.
The exact delineation when something counts as JS vs ASH is a little fuzzy,
since _technically_, the output of the `js` CLI command is an ASH
Value.

For the purposes of this PR, I'm using a very specific definition -- while
built-in library functions such as print / numeric_modifier are implemented
in Java, they use the ASH type system for representing inputs and outputs.
Indeed, we've seen problems with passing undefined / null to these
functions.

For the sake of clarity, also update the error message to make it clear
that the problem is converting into these ASH values.