Update dependencies. (#2413)
Most of these are only minor version bumps, but there were a handful of
major version changes.

Most notably: the maintainer for shadow has changed, along with the package
naming.

Hamcrest 3.0 looks like the main backwards incompatible change is that it's
ratcheted up the minimum Java version to require at least Java 8. Since
we're building with Java 17, this is a non-issue.

There were a handful of major version changes in dependencies that I'm not
sure we're ready for (flatlaf and swingx were the two that I recall), and a
couple of dependencies where the latest version is marked as beta or alpha
(spotless, slf4j-nop).