Throw JS exceptions when passing null / undefined to ASH. (#2381)
* Throw JS exceptions when passing null / undefined to ASH.

* Fix formatting.

* Change enforcement location.

The previous approach unfortunately had a bunch of false positives for
"void" functions (i.e. those that had no explicit return value), since the
implicit return value `undefined` triggered an exception when we tried to
convert that into an ASH or a Java value.

* Add missing .out files.

My .gitignore apparently ignored them...

* Throw exceptions for nested null / undefined.

It seems that NativeArray / NativeObject convert undefined -> null, so we
cannot differentiate at that level. It doesn't bother me too much. Both can
cause problems in ASH.

Note that in JavaScriptCommandTest, we cannot specify an object literal
directly -- this is one of the scenarios outlined in Gary Bernhardt's "wat"
talk, where the opening brace is parsed as the beginning of a block scope
and not the first character of an object literal. As such, those tests
begin with "o = " to provide the correct context to the interpreter.

* Remove now-redundant null checks.

We're already throwing an exception if either value or key is null
independently, so no point in checking this again.

* Test that null / undefined can be used in pure JS.

The exact delineation when something counts as JS vs ASH is a little fuzzy,
since _technically_, the output of the `js` CLI command is an ASH
Value.

For the purposes of this PR, I'm using a very specific definition -- while
built-in library functions such as print / numeric_modifier are implemented
in Java, they use the ASH type system for representing inputs and outputs.
Indeed, we've seen problems with passing undefined / null to these
functions.

For the sake of clarity, also update the error message to make it clear
that the problem is converting into these ASH values.

---------

Co-authored-by: Samuel Gaus <sam@gaus.co.uk>