feat: add modifier for 'lantern' (#2844)
Lanterns add extra spell damage equal to the largest hit of a spell, in a particular element (or elements).

Add two modifiers: "Lantern" giving the number of sources, where 1 is +100% damage; and "Lantern Element", giving the element(s) of the spell (omitted if nonconstant).

Porcelain porkpie is absent due to only working on PM spells. Novelty Belt Buckle is absent as the damage comes as an extra hit, and doesn't work for the wall of bones (and matters less for other monsters as it's only 10-20%).