Perform fuzzy matching for hermit CLI command.
I had tried using ItemFinder.getFirstMatchingItem(), but that apparently
tries to find matching items, independent of the values in the provided
source list.

Note that I removed 11-leaf clover from being added to hermitItems with a
count of -1 -- nothing seems to rely on this existing. However, the
presence of this entry was causing cloverCount() to return -1, which was
less than the 0 clovers needed.