Upgrade dependencies, gradlew.
Per dependencyUpdates, we have the following diffs remaining:

```
The following dependencies have later milestone versions:
 - com.diffplug.spotless:com.diffplug.spotless.gradle.plugin [6.23.3 -> 8.0.0]
 - com.formdev:flatlaf [1.6.5 -> 3.6.2]
     https://github.com/JFormDesigner/FlatLaf
 - com.formdev:flatlaf-intellij-themes [1.6.5 -> 3.6.2]
     https://github.com/JFormDesigner/FlatLaf
 - com.formdev:flatlaf-swingx [1.6.5 -> 3.6.2]
     https://github.com/JFormDesigner/FlatLaf
 - org.junit.jupiter:junit-jupiter [5.11.0 -> 6.0.0]
     https://junit.org/
 - org.junit.jupiter:junit-jupiter-api [5.11.0 -> 6.0.0]
     https://junit.org/
 - org.junit.jupiter:junit-jupiter-params [5.11.0 -> 6.0.0]
     https://junit.org/
 - org.junit.platform:junit-platform-launcher [1.11.0 -> 6.0.0]
     https://junit.org/
 - org.slf4j:slf4j-nop [2.0.17 -> 2.1.0-alpha1]
     http://www.slf4j.org
 - org.swinglabs:swingx [1.0 -> 1.6.1]
     http://www.swinglabs.org/

Gradle release-candidate updates:
 - Gradle: [9.1.0 -> 9.2.0-rc-2]
```

spotless 7.0+ introduces quite a few formatting diffs. If we choose to
accept these, that should be done in a standalone PR. It's not obvious that
(for instance) the treatment of switch expressions is desirable with
respect to continuations, but in general newer spotless versions will have
better support for newer Java features.

flatlaf, swingx continue to require legacy versions, as was the case in
https://github.com/kolmafia/kolmafia/pull/2413.

JUnit 6.0 seems to introduce a number of test breakages, possibly due to
changes in how the tests are run. This warrants further investigation.

Gradle, slf4j-nop have been upgraded to the latest stable (non-alpha,
non-release-candidate) version.

Nominal motivation is https://kolmafia.us/threads/31425.